+++ src/test/scala/sangria/schema/DefaultValuesSpec.scala
@@ -312,33 +312,34 @@ class DefaultValuesSpec extends WordSpec with Matchers with AwaitSupport {
       }
     }
 
-    "used with play JSON values" should {
-      "generated typeclass-based serialisation" in {
-        import play.api.libs.json._
-
-        implicit val sharesWrites = Json.writes[Shares]
-        implicit val commentWrites = Json.writes[Comment]
-
-        import sangria.integration.playJson._
-
-        val CustomInputType = complexInputType(
-          sharesDefault = Shares(123, 456),
-          commentsDefault = List(Comment("John Doe", "Nice post!", BigDecimal(100)), Comment("Foo", "Bar", BigDecimal(0.1))))
-
-        check(
-          CustomInputType,
-          defaultValue = Json.parse("""{"title": "Post #1", "text": "Amazing!"}"""),
-          expectedResult = Map(
-            "title" -> "Post #1",
-            "text" -> "Amazing!",
-            "tags" -> List("beginner", "scala"),
-            "views" -> 12,
-            "shares" -> Map("twitter" -> 123, "facebook" -> 456),
-            "comments" -> List(
-              Map("author" -> "John Doe", "text" -> "Nice post!", "likes" -> 100),
-              Map("author" -> "Foo", "text" -> "Bar", "likes" -> 0.1))),
-          expectedDefault = "{\"tags\":[\"beginner\",\"scala\"],\"text\":\"Amazing!\",\"shares\":{\"twitter\":123,\"facebook\":456},\"views\":12,\"title\":\"Post #1\",\"comments\":[{\"author\":\"John Doe\",\"text\":\"Nice post!\",\"likes\":100},{\"author\":\"Foo\",\"text\":\"Bar\",\"likes\":0.1}]}")
-      }
-    }
+    // FIXME: commented out because play requires Java 8 --> Move to integration library in order to remain compatible with Java 7 here
+//    "used with play JSON values" should {
+//      "generated typeclass-based serialisation" in {
+//        import play.api.libs.json._
+//
+//        implicit val sharesWrites = Json.writes[Shares]
+//        implicit val commentWrites = Json.writes[Comment]
+//
+//        import sangria.integration.playJson._
+//
+//        val CustomInputType = complexInputType(
+//          sharesDefault = Shares(123, 456),
+//          commentsDefault = List(Comment("John Doe", "Nice post!", BigDecimal(100)), Comment("Foo", "Bar", BigDecimal(0.1))))
+//
+//        check(
+//          CustomInputType,
+//          defaultValue = Json.parse("""{"title": "Post #1", "text": "Amazing!"}"""),
+//          expectedResult = Map(
+//            "title" -> "Post #1",
+//            "text" -> "Amazing!",
+//            "tags" -> List("beginner", "scala"),
+//            "views" -> 12,
+//            "shares" -> Map("twitter" -> 123, "facebook" -> 456),
+//            "comments" -> List(
+//              Map("author" -> "John Doe", "text" -> "Nice post!", "likes" -> 100),
+//              Map("author" -> "Foo", "text" -> "Bar", "likes" -> 0.1))),
+//          expectedDefault = "{\"tags\":[\"beginner\",\"scala\"],\"text\":\"Amazing!\",\"shares\":{\"twitter\":123,\"facebook\":456},\"views\":12,\"title\":\"Post #1\",\"comments\":[{\"author\":\"John Doe\",\"text\":\"Nice post!\",\"likes\":100},{\"author\":\"Foo\",\"text\":\"Bar\",\"likes\":0.1}]}")
+//      }
+//    }
   }
 }